feat: cerate check token

+++ src/services/api.ts
@@ -22,8 +22,10 @@ api.interceptors.response.use(
   },
   (error) => {
     if (error.response.status === 401) {
-      window.location.href = "/";
-      localStorage.clear();
+      if (window.location.pathname !== "/") {
+        window.location.href = "/";
+        localStorage.clear();
+      }
     }
 
     if (
